+++ bin/cli/src/config.rs
@@ -20,7 +20,7 @@ use anyhow::Context;
 use kailua_build::KAILUA_FPVM_ID;
 use kailua_common::config::config_hash;
 use kailua_contracts::SystemConfig;
-use kailua_host::fetch_rollup_config;
+use kailua_host::config::fetch_rollup_config;
 use risc0_zkvm::sha::Digest;
 
 #[derive(clap::Args, Debug, Clone)]
